resolving comment changes
